requirements.txt
pandas
numpy
spotipy
scipy
flask
python-dotenv
kagglehub